src/flowerui/requirements.txt
azure-identity
azure-storage-blob
azure-storage-queue
scikit-learn
requests
streamlit
